collapse e2e app tests into one suite, add localstorage mock in unit tests
