+++ tests/e2e/test_chaperone.js
@@ -16,6 +16,7 @@ log.info("Host Agent ID: %s", HOST_AGENT_ID);
 
 const REGISTERED_HAPP_HASH = "uhCkkCQHxC8aG3v3qwD_5Velo1IHE1RdxEr9-tuNSK15u73m1LPOo"
 const SUCCESSFUL_JOINING_CODE = Buffer.from(msgpack.encode('joining code')).toString('base64')
+const INVALID_JOINING_CODE = msgpack.encode('Failing joining Code').toString('base64')
 
 // Note: All envoyOpts.dnas will be registered via admin interface with the paths provided here
 const envoyOpts = {
@@ -60,6 +61,14 @@ describe("Server", () => {
     pageTestUtils.describeJsHandleLogs();
 
     await page.exposeFunction('delay', delay)
+    // must explicitly expose this native function to puppeteer, otherwise it is undefined and errors out
+    page.exposeFunction('showErrorMessage', error => {
+      if (error instanceof Error) {
+        log.silly(error.message)
+      } else {
+        log.silly(error)
+      }
+    })
 
     // Set logger settings for hosted app (in real word scenario - will be done when host installs app):
     try {
@@ -234,7 +243,7 @@ describe("Server", () => {
         await delay(15000);
         isSignedOut = await client.signOut();
         console.log('isSignedOut : ', isSignedOut)
-        console.log("Alice anonymous after sign-up: ", client.anonymous);
+        console.log("alice anonymous after sign-up: ", client.anonymous);
         const signedOutOnce = (isSignedOut && client.anonymous === true && client.agent_id !== "uhCAk6n7bFZ2_28kUYCDKmU8-2K9z3BzUH4exiyocxR6N5HvshouY")
 
         const isSignedIn = await client.signIn("alice.test.1@holo.host", "Passw0rd!");
@@ -247,7 +256,7 @@ describe("Server", () => {
         await delay(15000);
         isSignedOut = await client.signOut();
         console.log('isSignedOut : ', isSignedOut)
-        console.log("Alice anonymous after sign-in: ", client.anonymous);
+        console.log("alice anonymous after sign-in: ", client.anonymous);
         const signedOutTwice = (isSignedOut && client.anonymous === true && client.agent_id !== "uhCAk6n7bFZ2_28kUYCDKmU8-2K9z3BzUH4exiyocxR6N5HvshouY")
 
         return {
@@ -271,14 +280,94 @@ describe("Server", () => {
     }
   });
 
-  it("should sign-up on this Host");
-  
+  it("should sign-in with incorrect joining code and fail", async function() {
+    this.timeout(300_000);
+    const signupError = await page.evaluate(async function (host_agent_id, registered_happ_hash, invalidJoiningCode) {
+      console.log("Registered Happ Hash: %s", registered_happ_hash);
+      const client = new Chaperone({
+        "mode": Chaperone.DEVELOP,
+        "web_user_legend": {},
+        "connection": {
+          "ssl": false,
+          "host": "localhost",
+          "port": 4656,
+        },
+        host_agent_id, // used to assign host (id generated by hpos-seed)
+        app_id: registered_happ_hash, // NOT RANDOM: this needs to match the hash of app in hha
+        "timeout": 50000,
+        "debug": true,
+      });
+      client.skip_assign_host = true;
+
+      await client.ready(200_000);
+      let signupError
+      try {
+        // passing in a random/incorrect joining code
+        signupError = await client.signUp("carol.test.3@holo.host", "Passw0rd!", invalidJoiningCode);
+      } catch (error) {
+        console.log('Caught Sign-up Error: ', error)
+        return {
+          name: error.name,
+          message: error.message
+        }
+      }
+      console.log("Finished signed-up agent: %s", client.agent_id);
+      return signupError
+    }, HOST_AGENT_ID, REGISTERED_HAPP_HASH, INVALID_JOINING_CODE);
+
+    log.info("Completed evaluation: %s", signupError);
+    expect(signupError.name).to.equal('UserError');
+    expect(signupError.message).to.equal('Invalid joining code')
+  });
+
+  it("should sign-in with null joining code and fail", async function() {
+    this.timeout(300_000);
+    const signupError = await page.evaluate(async function (host_agent_id, registered_happ_hash) {
+      console.log("Registered Happ Hash: %s", registered_happ_hash);
+      const client = new Chaperone({
+        "mode": Chaperone.DEVELOP,
+        "web_user_legend": {},
+        "connection": {
+          "ssl": false,
+          "host": "localhost",
+          "port": 4656,
+        },
+        host_agent_id, // used to assign host (id generated by hpos-seed)
+        app_id: registered_happ_hash, // NOT RANDOM: this needs to match the hash of app in hha
+        "timeout": 50000,
+        "debug": true,
+      });
+      client.skip_assign_host = true;
+
+      await client.ready(200_000);
+      let signupError
+      try {
+        // passing in no joining code
+        signupError = await client.signUp("daniel.test.4@holo.host", "Passw0rd!", null);
+      } catch (error) {
+        console.log('Caught Sign-up Error: ', error)
+        return {
+          name: error.name,
+          message: error.message
+        }
+      }
+      console.log("Finished signed-up agent: %s", client.agent_id);
+      return signupError
+    }, HOST_AGENT_ID, REGISTERED_HAPP_HASH);
+
+    log.info("Completed evaluation: %s", signupError)
+    expect(signupError.name).to.equal('UserError')
+    expect(signupError.message).to.equal('Missing membrane proof')
+  })
+
+  it("should sign-up on this Host")
+
   it("should have no pending confirmations", async function() {
     // Give confirmation request some time to finish
-    this.timeout(5_000);
+    this.timeout(5_000)
     try {
       await delay(2_000);
       expect(envoy.pending_confirms).to.be.empty;
     } finally {}
-  });
-});
+  })
+})
