+++ tests/unit/test_server_mock_conductor.js
@@ -37,7 +37,6 @@ const envoyOpts = {
 }
 
 describe("Server with mock Conductor", () => {
-  const INTERNAL_INSTALLED_APP_ID = "holo-hosting-app"
   // Note: The value used for the hosted installed_app_ids
   // ** must match the hha_hash pased to the chaperone server (in setup_envoy.js)
   const HOSTED_INSTALLED_APP_ID = "uhCkkCQHxC8aG3v3qwD_5Velo1IHE1RdxEr9-tuNSK15u73m1LPOo"
@@ -84,7 +83,24 @@ describe("Server with mock Conductor", () => {
     await envoy.connected;
   });
   beforeEach('Set-up installed_app_ids for test', async () => {
-    appConductor.any(MOCK_CELL_DATA)
+    appConductor.any(MOCK_CELL_DATA);
+    // localstorage mock
+    const store = {};
+    const mockLocalStorage = {
+      getItem: function (key) {
+         return store[key]
+      },
+      setItem: function (key, value) {
+        return store[key] = value
+      },
+      removeItem: function (key) {
+        delete store[key]
+      },
+      clear: function () {
+        store = {}
+      },
+    }
+    global.window = { localStorage: mockLocalStorage }
   });
   afterEach("Close client", async function() {
     this.timeout(20_000)
